requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.18
streamlit>=1.28.0
plotly>=5.15.0
ta-lib>=0.4.25
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
pytz>=2023.3
